backend/requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.6.1
async-timeout==5.0.1
asyncpg==0.30.0
bcrypt==4.2.0
black==24.10.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.6
fastapi-cache2==0.2.2
fastapi-cli==0.0.5
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mypy-extensions==1.0.0
orjson==3.10.11
packaging==24.2
passlib==1.7.4
pathspec==0.12.1
pendulum==3.0.0
platformdirs==4.3.6
pydantic==2.9.2
pydantic-settings==2.7.1
pydantic_core==2.23.4
Pygments==2.18.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
pytz==2024.2
PyYAML==6.0.2
redis==4.6.0
rich==13.9.4
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
time-machine==2.16.0
typer==0.13.0
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.32.1
watchfiles==0.24.0
websockets==14.1
pypdf==5.4.0
reportlab==4.2.5
matplotlib==3.9.2
PyMuPDF==1.25.5
openpyxl==3.1.5
psycopg2-binary==2.9.10
numpy==1.26.4
pillow==11.0.0
textwrap3==0.9.2
setuptools
requests
pandas
